and fix compilation
