+++ bin/runtime-common/Cargo.toml
@@ -46,6 +46,7 @@ xcm-builder = { git = "https://github.com/paritytech/polkadot", branch = "master
 [dev-dependencies]
 bp-test-utils = { path = "../../primitives/test-utils" }
 pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "master" }
+pallet-bridge-messages = { path = "../../modules/messages", features = ["std", "test-helpers"] }
 
 [features]
 default = ["std"]
@@ -79,6 +80,7 @@ std = [
 runtime-benchmarks = [
 	"pallet-bridge-grandpa/runtime-benchmarks",
 	"pallet-bridge-messages/runtime-benchmarks",
+	"pallet-bridge-messages/test-helpers",
 	"pallet-bridge-parachains/runtime-benchmarks",
 	"pallet-bridge-relayers/runtime-benchmarks",
 	"xcm-builder/runtime-benchmarks",
