+++ bin/runtime-common/src/lib.rs
@@ -28,7 +28,6 @@ pub mod messages;
 pub mod messages_api;
 pub mod messages_benchmarking;
 pub mod messages_call_ext;
-pub mod messages_generation;
 pub mod messages_xcm_extension;
 pub mod parachains_benchmarking;
 pub mod priority_calculator;
